modal matches activities list
